# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI File Manager API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type FileItemType = typeof FileItemType[keyof typeof FileItemType];


export const FileItemType = {
  file: 'file',
  directory: 'directory',
} as const;

export interface FileItem {
  name: string;
  path: string;
  type: FileItemType;
  size: number;
  modifiedAt: string;
  mimeType?: string;
  isIndexed?: boolean;
}

export type SearchRequestMode = typeof SearchRequestMode[keyof typeof SearchRequestMode];


export const SearchRequestMode = {
  semantic: 'semantic',
  filename: 'filename',
  both: 'both',
} as const;

export interface SearchRequest {
  query: string;
  mode?: SearchRequestMode;
  path?: string;
  limit?: number;
}

export type SearchResultMatchType = typeof SearchResultMatchType[keyof typeof SearchResultMatchType];


export const SearchResultMatchType = {
  semantic: 'semantic',
  filename: 'filename',
  both: 'both',
} as const;

export interface SearchResult {
  file: FileItem;
  score: number;
  matchType: SearchResultMatchType;
  snippet?: string;
}

export type ActionAction = typeof ActionAction[keyof typeof ActionAction];


export const ActionAction = {
  rename: 'rename',
  delete: 'delete',
  move: 'move',
  copy: 'copy',
  create: 'create',
  code_execute: 'code_execute',
} as const;

export type ActionMode = typeof ActionMode[keyof typeof ActionMode];


export const ActionMode = {
  file_name: 'file_name',
  regex: 'regex',
  pattern: 'pattern',
  code: 'code',
} as const;

export type ActionParams = {
  replacement?: string;
  dest?: string;
  code?: string;
  content?: string;
  [key: string]: unknown;
 };

export type ActionReverseAction = { [key: string]: unknown };

export interface Action {
  id: string;
  action: ActionAction;
  mode: ActionMode;
  target: string;
  params?: ActionParams;
  isSafe: boolean;
  reverseAction?: ActionReverseAction;
}

export interface ActionPlan {
  transactionId: string;
  actionsSummary: string;
  actions: Action[];
  requiresCode: boolean;
  hasRiskyActions: boolean;
  estimatedTimeMs?: number;
}

export type PreviewResultPreviewFilesItem = {
  before?: string;
  after?: string;
  action: string;
};

export interface PreviewResult {
  transactionId: string;
  status: string;
  previewFiles: PreviewResultPreviewFilesItem[];
  errors?: string[];
}

export type TransactionStatus = typeof TransactionStatus[keyof typeof TransactionStatus];


export const TransactionStatus = {
  pending: 'pending',
  previewed: 'previewed',
  approved: 'approved',
  rejected: 'rejected',
  completed: 'completed',
  failed: 'failed',
} as const;

export interface Transaction {
  id: string;
  createdAt: string;
  updatedAt: string;
  status: TransactionStatus;
  command: string;
  actionsSummary: string;
  actions: Action[];
  snapshotId?: string;
  tmpPath?: string;
  completedActions?: string[];
}

export type SnapshotLocation = typeof SnapshotLocation[keyof typeof SnapshotLocation];


export const SnapshotLocation = {
  local: 'local',
  drive: 'drive',
} as const;

export interface Snapshot {
  id: string;
  timestamp: string;
  transactionId: string;
  affectedFiles: string[];
  location: SnapshotLocation;
  sizeMb: number;
  checksum?: string;
  restorable: boolean;
}

export interface CodeExecutionResult {
  success: boolean;
  output: string;
  error?: string;
  filesCreated?: string[];
  executionTimeMs: number;
}

export type ListFilesParams = {
path?: string;
};

export type ListFiles200 = {
  files: FileItem[];
  currentPath: string;
};

export type GetFileContentParams = {
path: string;
};

export type GetFileContent200 = {
  content: string;
  path: string;
  size: number;
  mimeType: string;
};

export type SearchFiles200 = {
  results: SearchResult[];
  query: string;
  mode: string;
};

export type IndexDirectoryBody = {
  path: string;
};

export type IndexDirectory200 = {
  status: string;
  filesIndexed: number;
};

export type ParseActionBody = {
  command: string;
  contextPath?: string;
};

export type PreviewActionsBody = {
  transactionId: string;
  actions: Action[];
};

export type ListTransactions200 = {
  transactions: Transaction[];
};

export type ApproveTransaction200 = {
  status: string;
  snapshotId?: string;
};

export type RejectTransaction200 = {
  status: string;
};

export type ListSnapshots200 = {
  snapshots: Snapshot[];
};

export type RestoreSnapshot200 = {
  status: string;
  filesRestored: number;
};

export type ExecuteCodeBody = {
  code: string;
  transactionId: string;
};